chore(ui): rename multiCheck folder to multi-check
