+++ packages/ui/src/components/index.ts
@@ -1,6 +1,6 @@
 export * from './button';
 export * from './icon';
 export * from './singleRadio';
-export * from './multiCheck';
+export * from './multi-check';
 export * from './checkbox';
 export * from './select';
